Additional commit

+++ backend/server.js
@@ -66,6 +66,10 @@ app.delete("/api/transactions/:id", async (req,res) => {
     try{
         const { id } = req.params
 
+        if(isNaN(parseInt(id))){
+            return res.status(400).json({ message: "Invalid transsaction ID" })
+        }
+
         const result = await sql`
             DELETE FROM transactions WHERE id = ${id} RETURNING * 
         `
